refactor(chat-server): extract auth into chat-server/auth.ts
Moves the Tailscale + bearer + trusted-proxy authentication logic
out of chat-server.ts into a focused module. No behavior change —
authenticateRequest() exits with the same {ok, identity, reason}
contract; warnIfAutoProxyTrust() preserves the startup warning that
TRUSTED_PROXY_IPS=auto headers aren't cryptographically verified.

chat-server.ts: -209 lines
chat-server/auth.ts: +207 lines (new)
Net wash, but the 207-line block is now self-contained and the
remaining chat-server.ts has one less concern to read past.
